VA3/requirements.txt
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.2.0
pkg-resources==0.0.0
